backend/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
chromadb==0.3.26
click==8.1.3
clickhouse-connect==0.6.14
coloredlogs==15.0.1
cryptography==41.0.1
data==0.4
dataclasses-json==0.5.9
decorator==5.1.1
Deprecated==1.2.14
dnspython==2.3.0
duckdb==0.9.1
exceptiongroup==1.1.1
fastapi==0.99.0
filelock==3.12.4
flatbuffers==23.5.26
fpdf==1.7.2
frozenlist==1.3.3
fsspec==2023.9.2
funcsigs==1.0.2
greenlet==2.0.2
h11==0.14.0
hnswlib==0.7.0
httptools==0.6.0
huggingface-hub==0.17.3
humanfriendly==10.0
idna==3.4
Jinja2==3.1.2
langchain==0.0.220
langchainplus-sdk==0.0.20
loguru==0.7.0
lz4==4.3.2
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.25.0
onnxruntime==1.16.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
overrides==7.4.0
packaging==23.1
pandas==2.1.1
pinecone-client==2.2.2
posthog==3.0.2
protobuf==4.24.4
pulsar-client==3.3.0
pycparser==2.21
pycryptodome==3.19.0
pydantic==1.10.10
PyGithub==1.59.0
PyJWT==2.7.0
PyNaCl==1.5.0
pypdf==3.11.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0
regex==2023.10.3
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.17
starlette==0.27.0
sympy==1.12
tenacity==8.2.2
tiktoken==0.4.0
tokenizers==0.14.1
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.7.0
tzdata==2023.3
urllib3==1.26.6
uvicorn==0.22.0
# uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
wrapt==1.15.0
yarl==1.9.2
zstandard==0.21.0
